requirements.txt
djangorestframework
channels
pypiwin32
channels_redis
django-cors-headers